pulling PostComponent into own file
